Update style
Bootstrap 4 was released so, I thought I'd give it a try. With this, I've had to update the style of the interface.

The interface is now fully dark. This means the navigation elements (mainly the sidebar and navbar) is now dark.

Labels have changed... Well, bootstrap have renamed them to "badges" so, I've been forced to call them the same.

I've re-done the sidebar so, it no longer uses lists. It's just plain 'ol links.

Alerts are now just a solid color instead of a gradient.. Ew, who likes gradients anyways?