Fixed the dynamic background in fullscreen mode
Fixed the issue of the background color being black after making it transparent to allow dynamic background color changes when not in fullscreen mode. This small JavaScript addition changes the background color when switching fullscreen mode and when switching map types. It works, and I've not noticed any performance changes myself, though it's probably not the prettiest way of doing this so I'm open for any changes.