Added python files used to extract PNGs from YTD files
A simple python script to turn YTD files into PNGs.
It literally goes thorugh the YTD archive and extracts ALL images it can find.
This is used to transform the minimap files "minimap_sea_*_*.ytd" to PNG files for use in the interface.

Just type `python extract_png.py` in the same directory as the YTD files.